+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Current (unreleased)
+
+- New filter options for areas: inside/approaching/both
+
 # 1.9.0 (2026-03-06)
 
 - Allow to filter observations per basis of record